[ci skip] Merge PR 21980
Merge PR #21980, commits were: 
 * include data sub-dir

The snpEff author hasn't made databases available online yet for v4.5, and has instead chosen to distribute SARS-CoV-2 annotations as part of the special covid19 release as if they had already been downloaded. For the bioconda package we usually omit the local data/ cache to save space since the databases are typically on sourceforge anyway and because everyone's needs are different. These changes copy the data/ sub-dir to the conda environment as a special case. If/when databases are available on sourceforge for v4_5 we should once again omit copying the data/ subdir.