[ci skip] Merge PR 22751
Merge PR #22751, commits were: 
 * Update rmats to build for python 2 and 3

Add conda_build_config.yaml for rmats to override python variants. The
intent is to build for both python 2 and 3. The bioconda
conda_build_config.yaml includes python 2 and 3, but it seems that the
bioconda variants do not override the conda-forge variants. Replacing
"*_cpython" with "*cpython" makes the strings unique and forces an
override.